backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
torch
faiss-cpu
sentence-transformers
numpy
requests
slowapi
PyPDF2

